Upgrades by Kenneth Verheggen :
- Relims now works both standalone and in cooperation with the automatic reprocessing system
- Updated the system to accept single projectIDs that are not in the relimsproperties file (runRelims method vs runClassicRelims method)
- Created a repository-option so every working projects has a folder with saved MGF and searchparameters file. Re_lims will check if there is a repository / folder with the project's 'sourcefiles' and run from there if possible. This greatly increases the speed of the search, as the searchparameters can be updated seperatly.
- Updated in a way that all projects have a seperate folder consisting out of a timestamp, projectprovider and projectID in a parentfolder that is user specific
- Made Re-Lims compatible with the new versions of peptideshaker and searchgui
- Created an errorguard  to deal with hanging/erroneous processes
- Implemented checkpoints (should be elaborated further)
- Created a file-manager to use a more generic approach in collecting the necessary files
- Implemented the use of a pride-xml converter (needs to be elaborated further in cooperation with pride-asap
- Fixed concurrency issues
- Solved minor performance bugs
- Made alternatives for previous "bad practice code" (for example, non-closing streams, uncaught exceptions)

+++ compomics-relims/src/main/java/com/compomics/relims/filemanager/FileGrabber.java
@@ -106,12 +106,14 @@ public class FileGrabber {
     }
 
     public static void deleteResultFolder() {
-        File resultFolder = RelimsProperties.getWorkSpace();
-        try {
-            FileUtils.deleteDirectory(resultFolder);
-        } catch (FileNotFoundException e) {
-        } catch (IOException ex) {
-            System.out.println("Could not delete the resultsfolder");
+        if (!RelimsVariableManager.getClassicMode()) {
+            File resultFolder = RelimsProperties.getWorkSpace();
+            try {
+                FileUtils.deleteDirectory(resultFolder);
+            } catch (FileNotFoundException e) {
+            } catch (IOException ex) {
+                System.out.println("Could not delete the resultsfolder");
+            }
         }
     }
 
@@ -120,6 +122,10 @@ public class FileGrabber {
         File dir = new File(RelimsProperties.getLocalPrideXMLRepository());
         FileFilter fileFilter = new WildcardFileFilter(wildCardFilterString);
         File[] files = dir.listFiles(fileFilter);
-        return files[0];
+        if (files != null) {
+            return files[0];
+        } else {
+            return null;
+        }
     }
 }
